// upsell-discount/generated/api.ts
export type Maybe<T> = T | null;
// ...
export type Attribute = {
  __typename?: 'Attribute';
  /** The key or name of the attribute. For example, `"customer_first_order"`. */
  key: Scalars['String']['output'];
  /** The value of the attribute. For example, `"true"`. */
  value?: Maybe<Scalars['String']['output']>;
};
// ...
export type Cart = {
  __typename?: 'Cart';
  /**
   * The custom attributes associated with a cart to store additional information. Cart attributes
   * allow you to collect specific information from customers on the **Cart** page, such as order notes,
   * gift wrapping requests, or custom product details. Attributes are stored as key-value pairs.
   */
  attribute?: Maybe<Attribute>;
  /**
   * Information about the customer that's interacting with the cart. It includes details such as the
   * customer's email and phone number, and the total amount of money the customer has spent in the store.
   * This information helps personalize the checkout experience and ensures that accurate pricing and delivery options
   * are displayed to customers.
   */
  buyerIdentity?: Maybe<BuyerIdentity>;
  /**
   * A breakdown of the costs that the customer will pay at checkout. It includes the total amount,
   * the subtotal before taxes and duties, the tax amount, and duty charges.
   */
  cost: CartCost;
  /** The items in a cart that are eligible for fulfillment and can be delivered to the customer. */
  deliverableLines: Array<DeliverableCartLine>;
  /**
   * A collection of items that are grouped by shared delivery characteristics. Delivery groups streamline
   * fulfillment by organizing items that can be shipped together, based on the customer's
   * shipping address. For example, if a customer orders a t-shirt and a pair of shoes that can be shipped
   * together, then the items are included in the same delivery group.
   *
   * In the [Order Discount](https://shopify.dev/docs/api/functions/reference/order-discounts) and
   * [Product Discount](https://shopify.dev/docs/api/functions/reference/product-discounts) legacy APIs,
   * the `cart.deliveryGroups` input is always an empty array. This means you can't access delivery groups when
   * creating Order Discount or Product Discount Functions. If you need to apply discounts to shipping costs,
   * then use the [Discount Function API](https://shopify.dev/docs/api/functions/reference/discount)
   * instead.
   */
  deliveryGroups: Array<CartDeliveryGroup>;
  /**
   * The items in a cart that the customer intends to purchase. A cart line is an entry in the
   * customer's cart that represents a single unit of a product variant. For example, if a customer adds two
   * different sizes of the same t-shirt to their cart, then each size is represented as a separate cart line.
   */
  lines: Array<CartLine>;
  /**
   * The additional fields on the **Cart** page that are required for international orders in specific countries,
   * such as customs information or tax identification numbers.
   */
  localizedFields: Array<LocalizedField>;
};
// ...
export type CartLinesDiscountsGenerateRunResult = {
  /** The list of operations to apply discounts to the cart. */
  operations: Array<CartOperation>;
};
// ...
export type Discount = HasMetafields & {
  __typename?: 'Discount';
  /** The [discount classes](https://shopify.dev/docs/apps/build/discounts/#discount-classes)) that the [discountNode](https://shopify.dev/docs/api/admin-graphql/latest/queries/discountNode)) supports. */
  discountClasses: Array<DiscountClass>;
  /**
   * A [custom field](https://shopify.dev/docs/apps/build/custom-data) that stores additional information
   * about a Shopify resource, such as products, orders, and
   * [many more](https://shopify.dev/docs/api/admin-graphql/latest/enums/MetafieldOwnerType).
   * Using [metafields with Shopify Functions](https://shopify.dev/docs/apps/build/functions/input-output/metafields-for-input-queries)
   * enables you to customize the checkout experience.
   */
  metafield?: Maybe<Metafield>;
};
// ...
export enum DiscountClass {
  /**
   * The discount is combined with an
   * [order discount](https://help.shopify.com/manual/discounts/combining-discounts/discount-combinations)
   * class.
   */
  Order = 'ORDER',
  /**
   * The discount is combined with a
   * [product discount](https://help.shopify.com/manual/discounts/combining-discounts/discount-combinations)
   * class.
   */
  Product = 'PRODUCT',
  /**
   * The discount is combined with a
   * [shipping discount](https://help.shopify.com/manual/discounts/combining-discounts/discount-combinations)
   * class.
   */
  Shipping = 'SHIPPING'
}
// ...
export type HasMetafields = {
  /**
   * A [custom field](https://shopify.dev/docs/apps/build/custom-data) that stores additional information
   * about a Shopify resource, such as products, orders, and
   * [many more](https://shopify.dev/docs/api/admin-graphql/latest/enums/MetafieldOwnerType).
   * Using [metafields with Shopify Functions](https://shopify.dev/docs/apps/build/functions/input-output/metafields-for-input-queries)
   * enables you to customize the checkout experience.
   */
  metafield?: Maybe<Metafield>;
};
// ...
export type Input = {
  __typename?: 'Input';
  /**
   * The cart where the Function is running. A cart contains the merchandise that a customer intends to purchase
   * and information about the customer, such as the customer's email address and phone number.
   */
  cart: Cart;
  /**
   * The discount node that owns the [Shopify
   * Function](https://shopify.dev/docs/apps/build/functions). Discounts are a way
   * for merchants to promote sales and special offers, or as customer loyalty
   * rewards. A single discount can be automatic or code-based, and can be applied
   * to a cart lines, orders, and delivery.
   */
  discount: Discount;
  /**
   * The discount codes that customers enter at checkout. Customers can enter codes
   * as an array of strings, excluding gift cards. Codes aren't validated in any
   * way other than to verify they aren't gift cards. This input is only available
   * in the `cart.lines.discounts.generate.fetch` and
   * `cart.delivery-options.discounts.generate.fetch` extension targets.
   */
  enteredDiscountCodes: Array<Scalars['String']['output']>;
  /**
   * The result of the fetch target. Refer to [network access](https://shopify.dev/apps/build/functions/input-output/network-access/graphql)
   * for Shopify Functions. This input is only available in the
   * `cart.lines.discounts.generate.run` and
   * `cart.delivery-options.discounts.generate.run` extension targets.
   */
  fetchResult?: Maybe<HttpResponse>;
  /**
   * The regional and language settings that determine how the Function
   * handles currency, numbers, dates, and other locale-specific values
   * during discount calculations. These settings are based on the store's configured
   * [localization practices](https://shopify.dev/docs/apps/build/functions/localization-practices-shopify-functions).
   */
  localization: Localization;
  /**
   * The exchange rate used to convert discounts between the shop's default
   * currency and the currency that displays to the customer during checkout.
   * For example, if a store operates in USD but a customer is viewing discounts in EUR,
   * then the presentment currency rate handles this conversion for accurate pricing.
   */
  presentmentCurrencyRate: Scalars['Decimal']['output'];
  /**
   * Information about the shop where the Function is running, including the shop's timezone
   * setting and associated [metafields](https://shopify.dev/docs/apps/build/custom-data).
   */
  shop: Shop;
  /**
   * The discount code entered by a customer, which caused the [Discount Function](https://shopify.dev/docs/apps/build/discounts#build-with-shopify-functions) to run.
   * This input is only available in the `cart.lines.discounts.generate.run` and
   * `cart.delivery-options.discounts.generate.run` extension targets.
   */
  triggeringDiscountCode?: Maybe<Scalars['String']['output']>;
};
// ...
export type Product = HasMetafields & {
  __typename?: 'Product';
  /**
   * A unique, human-readable string of the product's title. A handle can contain letters, hyphens (`-`), and
   * numbers, but not spaces. The handle is used in the online store URL for the product. For example, if a product
   * is titled "Black Sunglasses", then the handle is `black-sunglasses`.
   */
  handle: Scalars['Handle']['output'];
  /**
   * Whether the product is associated with any of the specified tags. The product must have at least one tag
   * from the list to return `true`.
   */
  hasAnyTag: Scalars['Boolean']['output'];
  /** Whether the product is associated with the specified tags. */
  hasTags: Array<HasTagResponse>;
  /**
   * A [globally-unique ID](https://shopify.dev/docs/api/usage/gids)
   * for the product.
   */
  id: Scalars['ID']['output'];
  /**
   * Whether the product is in any of the specified collections. The product must be in at least one collection
   * from the list to return `true`.
   *
   * A collection is a group of products that can be displayed in online stores and other sales channels in
   * categories, which makes it easy for customers to find them. For example, an athletics store might create
   * different collections for running attire and accessories.
   */
  inAnyCollection: Scalars['Boolean']['output'];
  /**
   * Whether the product is in the specified collections. The product must be in all of the collections in the
   * list to return `true`.
   *
   * A collection is a group of products that can be displayed in online stores and other sales channels in
   * categories, which makes it easy for customers to find them. For example, an athletics store might create
   * different collections for running attire and accessories.
   */
  inCollections: Array<CollectionMembership>;
  /** Whether the product is a gift card. */
  isGiftCard: Scalars['Boolean']['output'];
  /**
   * A [custom field](https://shopify.dev/docs/apps/build/custom-data) that stores additional information
   * about a Shopify resource, such as products, orders, and
   * [many more](https://shopify.dev/docs/api/admin-graphql/latest/enums/MetafieldOwnerType).
   * Using [metafields with Shopify Functions](https://shopify.dev/docs/apps/build/functions/input-output/metafields-for-input-queries)
   * enables you to customize the checkout experience.
   */
  metafield?: Maybe<Metafield>;
  /**
   * A custom category for a product. Product types allow merchants to define categories other than the
   * ones available in Shopify's
   * [standard product categories](https://help.shopify.com/manual/products/details/product-type).
   */
  productType?: Maybe<Scalars['String']['output']>;
  /**
   * The localized name for the product that displays to customers. The title is used to construct the product's
   * handle, which is a unique, human-readable string of the product's title. For example, if a product is titled
   * "Black Sunglasses", then the handle is `black-sunglasses`.
   */
  title: Scalars['String']['output'];
  /** The name of the product's vendor. */
  vendor?: Maybe<Scalars['String']['output']>;
};
// ...
export enum ProductDiscountSelectionStrategy {
  /** Apply all the discount candidates to eligible cart lines. This doesn't override discount combination or stacking rules. */
  All = 'ALL',
  /** Apply the first discount candidate to cart lines that satisfies conditions. */
  First = 'FIRST',
  /** Apply the discount to the cart line that offers the maximum reduction. */
  Maximum = 'MAXIMUM'
}
// ...
export type CartInput = { __typename?: 'Input', cart: { __typename?: 'Cart', attribute?: { __typename?: 'Attribute', value?: string | null } | null, lines: Array<{ __typename?: 'CartLine', id: string, upsellAttribute?: { __typename?: 'Attribute', value?: string | null } | null, bundleAttribute?: { __typename?: 'Attribute', value?: string | null } | null }> }, discount: { __typename?: 'Discount', discountClasses: Array<DiscountClass> } };